+++ reimi/frontend/reimi_app/lib/application/usecases/chat_room/get_chat_room_summaries_usecase.dart
@@ -1,11 +1,11 @@
-import 'package:reimi_app/data/models/chat_room_summary_model.dart';
+import 'package:reimi_app/domain/read_models/chat_room_summary_read_model.dart';
 import 'package:reimi_app/domain/repositories/chat_room_repository.dart';
 
 class GetChatRoomSummariesUsecase {
   const GetChatRoomSummariesUsecase(this._repository);
   final ChatRoomRepository _repository;
 
-  Future<List<ChatRoomSummaryModel>?> call(String userId) {
-    return _repository.fetchChatRoomSummaries(userId);
+  Future<List<ChatRoomSummaryReadModel>> call() {
+    return _repository.fetchChatRoomSummaries();
   }
 }
