+++ reimi/frontend/reimi_app/lib/application/usecases/chat_room/get_unmessaged_match_users_usecase.dart
@@ -1,11 +1,11 @@
-import 'package:reimi_app/data/models/unmessaged_match_user_model.dart';
+import 'package:reimi_app/domain/read_models/unmessaged_match_user_read_model.dart';
 import 'package:reimi_app/domain/repositories/chat_room_repository.dart';
 
 class GetUnmessagedMatchUsersUseCase {
   const GetUnmessagedMatchUsersUseCase(this._repository);
   final ChatRoomRepository _repository;
 
-  Future<List<UnmessagedMatchUserModel>?> call(String userId) {
-    return _repository.fetchUnmessagedMatchUsers(userId);
+  Future<List<UnmessagedMatchUserReadModel>> call() {
+    return _repository.fetchUnmessagedMatchUsers();
   }
 }
