--- reimi/frontend/reimi_app/lib/domain/usecases/like/get_like_users_to_user_usecase.dart
@@ -1,15 +1,0 @@
-import 'package:reimi_app/data/models/like_user_model.dart';
-import 'package:reimi_app/domain/repositories/like_repository.dart';
-
-class GetLikeUsersToUserUseCase {
-  const GetLikeUsersToUserUseCase(this._repository);
-  final LikeRepository _repository;
-
-  Future<List<LikeUserModel>?> call(String userId) async {
-    final likeUsers = await _repository.getLikeUsersToUser(userId);
-    if (likeUsers == null) {
-      return null;
-    }
-    return likeUsers;
-  }
-}
